+++ package.json
@@ -16,6 +16,7 @@
     "axios": "^1.4.0",
     "firebase": "^10.1.0",
     "localforage": "^1.10.0",
+    "lottie-react": "^2.4.0",
     "match-sorter": "^6.3.1",
     "react": "^18.2.0",
     "react-card-slider-component": "^1.0.9",
