+++ src/pages/Login-&-singup/Login.css
@@ -1,0 +1,7 @@
+.login-singUp-bg {
+  background-image: url("../../assets/images/login-images/login-bg.png");
+  background-size: contain;
+  background-repeat: no-repeat;
+  background-attachment: fixed;
+  height: 100vh;
+}
