+++ src/pages/Login-&-singup/Login.jsx
@@ -1,0 +1,129 @@
+import Lottie from "lottie-react";
+import { useState } from "react";
+import { useForm } from "react-hook-form";
+import { AiFillEye, AiFillEyeInvisible } from "react-icons/ai";
+import { ImWarning } from "react-icons/im";
+import { Link, useLocation, useNavigate } from "react-router-dom";
+import Swal from "sweetalert2";
+import Logo from "../../assets/Logo/logo.svg";
+import loginAnimation from "../../assets/images/login-images/login.json";
+import useAuth from "../../hooks/useAuth";
+import { addUser } from "../../hooks/userApi";
+import "./Login.css";
+import SocialSigning from "./SocialSigning";
+
+const Login = () => {
+  const [error, setError] = useState("");
+  const [showPassword, setShowPassword] = useState(false);
+  const location = useLocation();
+
+  const { signIn } = useAuth();
+  const navigate = useNavigate();
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm();
+  let from = location.state?.from?.pathname || "/";
+
+  const handleTogglePassword = () => {
+    setShowPassword(!showPassword);
+  };
+
+  const onSubmit = (data) => {
+    signIn(data.email, data.password)
+      .then((result) => {
+        if (result.user) {
+          Swal.fire({
+            icon: "success",
+            title: "Your LogIn Successfully",
+            showConfirmButton: false,
+            timer: 2500,
+          });
+          addUser(result.user);
+        }
+        setError("");
+        navigate(from, { replace: true });
+      })
+      .catch((err) => {
+        setError(err.message);
+      });
+  };
+
+  return (
+    <div className="login-singUp-bg ">
+      <div className="nav-container items-center grid grid-cols-1 md:grid-cols-2 ">
+        <div className="w-full">
+          <Lottie animationData={loginAnimation} loop />
+        </div>
+        <div className="card w-full ">
+          <div className="mx-auto mb-5">
+            <Link to="/">
+              <img className=" w-28" src={Logo} alt="logo" />
+            </Link>
+          </div>
+          <form onSubmit={handleSubmit(onSubmit)} className="">
+            <div className="form-control">
+              <label className="label">
+                <span className="label-text">Email</span>
+              </label>
+              <input type="email" placeholder="email" className="input border-gray-3 w-full focus:input-bordered input-accent" {...register("email", { required: true })} />
+              {errors.email && (
+                <p className="text-red-600">
+                  <ImWarning className="inline-block" /> Email is required
+                </p>
+              )}
+            </div>
+            <div className="form-control">
+              <label className="label">
+                <span className="label-text mt-3">Password</span>
+              </label>
+              <div className="relative">
+                <input
+                  type={showPassword ? "text" : "password"}
+                  placeholder="Enter your password"
+                  className="input border-gray-3 w-full focus:input-bordered input-accent"
+                  {...register("password", { required: true })}
+                />
+                <button type="button" onClick={handleTogglePassword} className="absolute right-5 bottom-4 text-my-primary">
+                  {showPassword ? <AiFillEye /> : <AiFillEyeInvisible />}
+                </button>
+              </div>
+              {errors.password && (
+                <p className="text-red-600">
+                  <ImWarning className="inline-block" /> password is required
+                </p>
+              )}
+            </div>
+            <button type="button" className="text-xs hover:underline mt-2">
+              Forgot password?
+            </button>
+            <label className="label">
+              {error && (
+                <p className="text-red-600">
+                  <ImWarning className="inline-block" /> {error}
+                </p>
+              )}
+            </label>
+
+            <button type="submit" className="my-btn w-full mx-auto form-control mt-2">
+              LogIn
+            </button>
+          </form>
+          <div className="divider">OR</div>
+          <SocialSigning />
+          <label className="">
+            <p className=" mt-2">
+              <span>Don&apos;t Have An Account ?</span>{" "}
+              <Link to="/signUp" className="underline font-semibold  hover:text-my-primary">
+                Sign Up
+              </Link>
+            </p>
+          </label>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Login;
