+++ src/routes/Routes.jsx
@@ -26,17 +26,17 @@ import ViewCart from "../pages/Dashboard/UserDashboard/ViewCart/ViewCart";
 import HealthTips from "../pages/HealthTips/HealthTips";
 import HealthTipsDetails from "../pages/HealthTipsDetails/HealthTipsDetails";
 import Home from "../pages/Home/Home/Home";
-import Login from "../pages/Login/Login";
+import MedicineCarts from "../pages/MedicineCarts/MedicineCarts";
 import MedicineDetails from "../pages/Medicines/MedicineDetails";
 import Medicines from "../pages/Medicines/Medicines";
 import PharmacyRegistrationPage from "../pages/PharmacyRegistrationPage/PharmacyRegistrationPage";
+import Services from "../pages/Services/Services";
 import ErrorPage from "../pages/Shared/ErrorPage/ErrorPage";
 import LabPayment from "../pages/Shared/LabTest/LabPayment";
 import LabTest from "../pages/Shared/LabTest/LabTest";
 import LabTestPage from "../pages/Shared/LabTest/LabTestPage";
-import SignUp from "../pages/SignUp/SignUp";
-import MedicineCarts from "../pages/MedicineCarts/MedicineCarts";
-import Services from "../pages/Services/Services";
+import Login from "../pages/Login-&-singup/Login";
+import SignUp from "../pages/Login-&-singup/SignUp";
 
 const router = createBrowserRouter([
   {
